+++ backend/.env.example
@@ -27,10 +27,10 @@ DB_DATABASE=laravel
 DB_USERNAME=laravel
 DB_PASSWORD=password
 
-SESSION_DRIVER=database
+SESSION_DRIVER=file
 SESSION_LIFETIME=120
-SESSION_ENCRYPT=false
-SESSION_PATH=/
+SESSION_ENCRYPT=true
+SESSION_PATH=storage/framework/sessions
 SESSION_DOMAIN=null
 
 BROADCAST_CONNECTION=log
