+++ backend/app/Http/Controllers/Api/DeckController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Models\Deck;
 use App\Http\Requests\DeckRequest;
 use Illuminate\Http\Request;
